requirements.txt
Django
bs4
diceroller
discord.py
lxml
pandas
